requirements.txt
streamlit==1.32.0
pandas==2.2.0
plotly==5.18.0
langchain
langchain_community
unstructured
libmagic
python-magic
python-magic-bin
faiss-cpu
sentence-transformers
streamlit
python-dotenv
langchain_ollama
langchain_huggingface
sentence_transformers
openpyxl
scikit-learn